src/requirements.txt
fastapi==0.65.2
pydantic==1.10.15
uvicorn==0.13.4
python-dateutil==2.8.1
